src/main/java/com/crud/library/mapper/BookMapper.java
package com.crud.library.mapper;

import com.crud.library.domain.Book;
import com.crud.library.dto.BookDto;
import org.springframework.stereotype.Component;

import java.util.ArrayList;
import java.util.List;
import java.util.stream.Collectors;

@Component
public class BookMapper {

    public BookDto mapToBookDto(Book book) {
        return new BookDto(
                book.getId(),
                book.getTitle(),
                book.getAuthor(),
                book.getPubYear());
    }

    public Book mapToBook(BookDto bookDto) {
        return new Book(
                bookDto.getId(),
                bookDto.getTitle(),
                bookDto.getAuthor(),
                bookDto.getPubYear(),
                new ArrayList<>());
    }

    public List<BookDto> mapToBookDtos(List<Book> books) {
        return books.stream()
                .map(this::mapToBookDto)
                .collect(Collectors.toList());
    }

    public List<Book> mapToBooks(List<BookDto> bookDtos) {
        return bookDtos.stream()
                .map(this::mapToBook)
                .collect(Collectors.toList());
    }
}
